# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aura API
  version: 0.1.0
  description: REST API for Aura sports medicine integrations.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/athletes/{id}/score:
    post:
      summary: Recalculate an athlete injury risk score
      tags: [Score]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Score recalculated
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                  recommendations:
                    type: array
                    items:
                      type: object
                  recommendationLogId:
                    type: [string, 'null']
                  recommendationLogError:
                    type: [string, 'null']
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/rehab/{sessionId}/rtp:
    post:
      summary: Update RTP criteria for a rehab session
      tags: [RTP]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rtp_criteria]
              properties:
                rtp_criteria:
                  type: array
                  items:
                    $ref: '#/components/schemas/RtpCriterion'
      responses:
        '200':
          description: RTP criteria updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, invalid, or expired Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RtpCriterion:
      type: object
      required: [label, done]
      properties:
        label:
          type: string
        done:
          type: boolean
...